+++ sdk/appconfiguration/azure-data-appconfiguration/src/samples/java/com/azure/data/appconfiguration/EntraIdAuthentication.java
@@ -3,6 +3,7 @@
 
 package com.azure.data.appconfiguration;
 
+import com.azure.core.util.Configuration;
 import com.azure.data.appconfiguration.models.ConfigurationSetting;
 import com.azure.identity.DefaultAzureCredential;
 import com.azure.identity.DefaultAzureCredentialBuilder;
@@ -10,7 +11,7 @@ import com.azure.identity.DefaultAzureCredentialBuilder;
 /**
  * Sample demonstrates how to use AAD token to build a configuration client.
  */
-public class AadAuthentication {
+public class EntraIdAuthentication {
     /**
      * Sample for how to use AAD token Authentication.
      *
@@ -19,7 +20,7 @@ public class AadAuthentication {
     public static void main(String[] args) {
         // The endpoint can be obtained by going to your App Configuration instance in the Azure portal
         // and navigating to "Overview" page. Looking for the "Endpoint" keyword.
-        String endpoint = "{endpoint_value}";
+        String endpoint = Configuration.getGlobalConfiguration().get("AZ_CONFIG_ENDPOINT");
 
         // Default token credential could be obtained from Identity service.
         // It tries to create a valid credential in the following order: